chore: split build into fine-grained modules
