+++ src/index.ts
@@ -1,4 +1,4 @@
-import build, { BuildCliFlags } from './lib/build';
-import preview, { PreviewOption } from './lib/preview';
+import build, { BuildCliFlags } from './commands/build';
+import preview, { PreviewOption } from './commands/preview';
 
 export { build, preview, BuildCliFlags, PreviewOption };
