+++ src/pdf.ts
@@ -1,0 +1,185 @@
+import fs from 'fs';
+import path from 'path';
+import chalk from 'chalk';
+import puppeteer from 'puppeteer';
+import { Meta, Payload, TOCItem } from './broker';
+import { PostProcess } from './postprocess';
+import {
+  getBrokerUrl,
+  launchSourceAndBrokerServer,
+  LoadMode,
+  PageSize,
+} from './server';
+import {
+  log,
+  statFile,
+  findEntryPointFile,
+  debug,
+  launchBrowser,
+} from './util';
+
+export async function buildPDF({
+  input,
+  rootDir,
+  outputPath,
+  size,
+  loadMode,
+  executableChromium,
+  sandbox,
+  verbose,
+  timeout,
+  pressReady,
+}: {
+  input: string;
+  rootDir: string | undefined;
+  outputPath: string;
+  size: string | number | undefined;
+  loadMode: LoadMode;
+  executableChromium: string | undefined;
+  sandbox: boolean;
+  verbose: boolean;
+  timeout: number;
+  pressReady: boolean;
+}) {
+  const stat = await statFile(input);
+  const root = rootDir || (stat.isDirectory() ? input : path.dirname(input));
+  debug(root);
+  const sourceIndex = await findEntryPointFile(input, root);
+  debug('index', sourceIndex);
+
+  const outputFile =
+    fs.existsSync(outputPath) && fs.statSync(outputPath).isDirectory()
+      ? path.resolve(outputPath, 'output.pdf')
+      : outputPath;
+  const outputSize = size ? parsePageSize(size) : undefined;
+
+  const [source, broker] = await launchSourceAndBrokerServer(root);
+  const sourcePort = source.port;
+  const brokerPort = broker.port;
+  debug(source, broker);
+  const navigateURL = getBrokerUrl({
+    sourcePort,
+    sourceIndex,
+    brokerPort,
+    loadMode,
+    outputSize,
+  });
+  debug('brokerURL', navigateURL);
+
+  log(`Launching build environment...`);
+  debug(
+    `Executing Chromium path: ${
+      executableChromium || puppeteer.executablePath()
+    }`,
+  );
+  const browser = await launchBrowser({
+    headless: true,
+    executablePath: executableChromium || puppeteer.executablePath(),
+    // Why `--no-sandbox` flag? Running Chrome as root without --no-sandbox is not supported. See https://crbug.com/638180.
+    args: [sandbox ? '' : '--no-sandbox'],
+  });
+  const version = await browser.version();
+  debug(chalk.green('success'), `version=${version}`);
+
+  const page = await browser.newPage();
+
+  page.on('pageerror', (error) => {
+    log(chalk.red(error.message));
+  });
+
+  page.on('console', (msg) => {
+    if (/time slice/.test(msg.text())) return;
+    if (!verbose) return;
+    log(chalk.gray(msg.text()));
+  });
+
+  page.on('response', (response) => {
+    debug(
+      chalk.gray('broker:response'),
+      chalk.green(response.status().toString()),
+      response.url(),
+    );
+    if (300 > response.status() && 200 <= response.status()) return;
+    log(chalk.red(`${response.status()}`, response.url()));
+  });
+
+  log('Building pages...');
+
+  await page.goto(navigateURL, { waitUntil: 'networkidle0' });
+  await page.waitFor(() => !!window.coreViewer);
+
+  const metadata = await loadMetadata(page);
+  const toc = await loadTOC(page);
+
+  await page.emulateMediaType('print');
+  await page.waitForFunction(
+    () => window.coreViewer.readyState === 'complete',
+    {
+      polling: 1000,
+      timeout,
+    },
+  );
+
+  log('Generating PDF...');
+
+  const pdf = await page.pdf({
+    margin: {
+      top: 0,
+      bottom: 0,
+      right: 0,
+      left: 0,
+    },
+    printBackground: true,
+    preferCSSPageSize: true,
+  });
+
+  log('Processing PDF...');
+
+  await browser.close();
+
+  const post = await PostProcess.load(pdf);
+  await post.metadata(metadata);
+  await post.toc(toc);
+  await post.save(outputFile, { pressReady });
+  return outputFile;
+}
+function parsePageSize(size: string | number): PageSize {
+  const [width, height, ...others] = size ? `${size}`.split(',') : [];
+  if (others.length) {
+    throw new Error(`Cannot parse size: ${size}`);
+  } else if (width && height) {
+    return {
+      width,
+      height,
+    };
+  } else {
+    return {
+      format: width || 'Letter',
+    };
+  }
+}
+
+async function loadMetadata(page: puppeteer.Page): Promise<Meta> {
+  return page.evaluate(() => window.coreViewer.getMetadata());
+}
+// Show and hide the TOC in order to read its contents.
+// Chromium needs to see the TOC links in the DOM to add
+// the PDF destinations used during postprocessing.
+
+async function loadTOC(page: puppeteer.Page): Promise<TOCItem[]> {
+  return page.evaluate(
+    () =>
+      new Promise<TOCItem[]>((resolve) => {
+        function listener(payload: Payload) {
+          if (payload.a !== 'toc') {
+            return;
+          }
+          window.coreViewer.removeListener('done', listener);
+          window.coreViewer.showTOC(false);
+          resolve(window.coreViewer.getTOC());
+        }
+        window.coreViewer.addListener('done', listener);
+        window.coreViewer.showTOC(true);
+      }),
+  );
+}
